remove duplicate CAT functionality #36
