+++ setup.py
@@ -41,7 +41,6 @@ setup(
     ],
     entry_points={
         'console_scripts': [
-            'cosmo=swan.cosmo.cosmo:main',
             'modeller=swan.models.modeller:main',
             'save_dataset_in_db=swan.data.save_data:main',
             'smiles_screener=swan.filter.screen:main'
@@ -49,7 +48,6 @@ setup(
     },
     install_requires=[
         'CAT@git+https://github.com/nlesc-nano/CAT@master',
-        'nano-CAT@git+https://github.com/nlesc-nano/nano-CAT@master',
         'data-CAT@git+https://github.com/nlesc-nano/data-CAT@master',
         'horovod', 'mendeleev', 'more_itertools', 'numpy', 'pandas',
         'pyyaml>=5.1.1', 'retry', 'seaborn', 'schema', 'sqlalchemy',
