--- swan/cosmo/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-from .cat_interface import call_mopac
-from .functions import merge_csv
-
-
-__all__ = ["call_mopac", "merge_csv"]
