+++ tests/test_auxiliar_functions.py
@@ -1,5 +1,4 @@
-from swan.cosmo.functions import merge_csv, run_command
-from pathlib import Path
+from swan.cat_interface import run_command
 
 
 def test_run_command():
@@ -9,13 +8,3 @@ def test_run_command():
     rs, _ = run_command(cmd)
 
     assert rs.split()[0].decode() == '20'
-
-
-def test_merge_csv(tmp_path):
-    """Check the merge of csv files containing the Gammas."""
-    path_files = "tests/test_files"
-    output = Path(tmp_path) / "merged.csv"
-
-    df = merge_csv(path_files, output)
-
-    assert 'gammas' in df.columns
